+++ CHANGELOG.md
@@ -18,4 +18,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `removeAccents`
     - `join`
     - `countWords`
-    - `isNumber`
+    - `isNumber` | `isNotNumber`